[MOVE] : SeatGrade 파일 위치 이동
- SeatGrade 도 엔티티이므로 독립적인 패키지를 가지게 하였음.
