--- src/main/java/com/prgrms/be/intermark/domain/performance/dto/PerformanceCommandResponseDTO.java
@@ -1,7 +1,0 @@
-package com.prgrms.be.intermark.domain.performance.dto;
-
-import lombok.Builder;
-
-@Builder
-public record PerformanceCommandResponseDTO(Long performanceId) {
-}
