+++ src/main/java/com/prgrms/be/intermark/domain/schedule_seat/model/ScheduleSeat.java
@@ -1,17 +1,29 @@
 package com.prgrms.be.intermark.domain.schedule_seat.model;
 
+import java.util.Objects;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.UniqueConstraint;
+import javax.validation.constraints.NotNull;
+
+import org.springframework.util.Assert;
+
 import com.prgrms.be.intermark.domain.schedule.model.Schedule;
 import com.prgrms.be.intermark.domain.seat.model.Seat;
-import com.prgrms.be.intermark.domain.seat.model.SeatGrade;
+import com.prgrms.be.intermark.domain.seatgrade.model.SeatGrade;
+
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.util.Assert;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import java.util.Objects;
 
 @Entity
 @Table(name = "schedule_seat",
